dmaengine: qcom: bam_dma: Add remote power collapse mode
In some configurations, we are required to configure the BAM,
but the BAM clk is managed by a remote processor. It is only turned
on after coordination with the remote processor, and while turned off
accessing the registers results in a bus stall.

Attempt to support this mode by only accessing BAM registers when
a DMA channel is actually requested. Otherwise, keep the BAM in reset.