arm64: dts: qcom: msm8916-samsung-e7: Add initial dts
This dts adds support for Samsung Galaxy E7 smartphone released in 2015.

Add a device tree for E7 with initial support for:

- GPIO keys
- SDHCI (internal and external storage)
- USB Device Mode
- UART
- WCNSS (WiFi/BT)
- Regulators

Unfortunately, some variants were released with outdated 32-bit only
firmware and never received any update from Samsung. Since the 32-bit
TrustZone firmware is signed there seems to be no way currently to
actually boot this device tree on arm64 Linux at the moment. :(

However, it is possible to use this device tree by compiling an ARM32 kernel
instead. The device tree can be easily built on ARM32 with an #include
and it works really well there. To avoid confusion for others it is still
better to add this device tree on arm64. Otherwise it's easy to forget
to update this one when making some changes that affect all MSM8916 devices.

Maybe someone finds a way to boot ARM64 Linux on 32-bit devices at some point.
In this case I expect that this device tree can be simply used as-is.

Signed-off-by: Lin, Meng-Bo <linmengbo0689@protonmail.com>