arm64: dts: qcom: msm8916-samsung-e7: Add initial dts
Samsung Galaxy E7 is a smartphone using the MSM8916 SoC released in 2015.

The E7 and A3/A5 are similar, with some differences in accelerometer,
MUIC, Vibator, WCNSS and touch key. The common parts are shared in
msm8916-samsung-a2015-common.dtsi to reduce duplication.

Motor is powered by VPH_PWR (aka Vbat) and controlled by GPIO.

The difference of accelerometer between E7/A3/A5 is the sensor chip.
ST LIS2HH12 is used instead.

On E7 the touch key is supplied by a single fixed regulator (enabled via
GPIO 97) that supplies both MCU and LED.

Unfortunately, some E7 were released with outdated 32-bit only firmware
and never received any update from Samsung. Since the 32-bit TrustZone
firmware is signed there seems to be no way currently to actually boot
this device tree on arm64 Linux on those variants at the moment. :(

However, it is possible to use this device tree by compiling an ARM32
kernel instead. The device tree can be easily built on ARM32 with
an #include and it works really well there. To avoid confusion for others
it is still better to add this device tree on arm64. Otherwise it's easy
to forget to update this one when making some changes that affect all
MSM8916 devices.

Maybe someone finds a way to boot ARM64 Linux on those device at some
point. In this case I expect that this device tree can be simply used
as-is.

Signed-off-by: Lin, Meng-Bo <linmengbo0689@protonmail.com>
Co-developed-by: Stephan Gerhold <stephan@gerhold.net>
Signed-off-by: Stephan Gerhold <stephan@gerhold.net>