arm64: dts: qcom: msm8916-samsung-j3ltetw: Add NFC
This version of the J3 2016 has a NXP PN547, which is supported
by the nxp-nci-i2c driver in mainline. It seems to detect NFC tags
using "nfctool" just fine, although more testing is difficult given
there seem to be very few useful applications making use of the
Linux NFC subsystem. :(

Add NXP PN547 together with a necessary fixed regulator pvdd-supply
for it.

Signed-off-by: Lin, Meng-Bo <linmengbo0689@protonmail.com>