arm64: dts: qcom: msm8916-samsung-gprime: Add initial device tree (v5)
v5: Add GPIO buttons, WCNSS and arm32 device tree

The Samsung Galaxy Grand Prime is a phone based on MSM8916. It is similar
to the other Samsung devices based on MSM8916 with only a few minor
differences.

The device tree contains initial support for the Grand Prime with:
 - UART
 - eMMC/SD card
 - Buttons
 - WiFi/Bluetooth (WCNSS)
 - USB

Unfortunately, some Grand Prime were released with outdated 32-bit only
firmware and never received any update from Samsung. Since the 32-bit
TrustZone firmware is signed there seems to be no way currently to
actually boot this device tree on arm64 Linux on those variants at the
moment. :(

However, it is possible to use this device tree by compiling an ARM32
kernel instead. The device tree can be easily built on ARM32 with
an #include and it works really well there. To avoid confusion for others
it is still better to add this device tree on arm64. Otherwise it's easy
to forget to update this one when making some changes that affect all
MSM8916 devices.

Maybe someone finds a way to boot ARM64 Linux on those device at some
point. In this case I expect that this device tree can be simply used
as-is.

Signed-off-by: Lin, Meng-Bo <linmengbo0689@protonmail.com>