arm64: dts: qcom: msm8916-samsung-grandmax: Add initial dts
Samsung Galaxy Grand Max is a smartphone using the MSM8916 SoC released in
2015.

Grand Max and E5/E7 are similar, with some differences in hall sensor,
panel, touchscreen and touchkey. The common parts are shared in
msm8916-samsung-e2015-common.dtsi to reduce duplication.

Unfortunately, some Grand Max were released with outdated 32-bit only firmware
and never received any update from Samsung. Since the 32-bit TrustZone
firmware is signed there seems to be no way currently to actually boot
this device tree on arm64 Linux on those variants at the moment. :(

However, it is possible to use this device tree by compiling an ARM32
kernel instead. The device tree can be easily built on ARM32 with
an #include and it works really well there. To avoid confusion for others
it is still better to add this device tree on arm64. Otherwise it's easy
to forget to update this one when making some changes that affect all
MSM8916 devices.

Maybe someone finds a way to boot ARM64 Linux on those device at some
point. In this case I expect that this device tree can be simply used
as-is.

Signed-off-by: Lin, Meng-Bo <linmengbo0689@protonmail.com>