Input: synaptics-rmi4 - workaround touchscreens with no pressure
Some Synaptics touchscreens with F12 do not seem to report pressure (z)
information. Unfortunately, in this case the "z" field is missing in
the data, causing the width and all further objects to be shifted by
one byte. Also, there seems to be no way to detect this dynamically
because the RMI item register description still says there are 8 bytes
for each finger/object, even though there are only 7 in this case.

Make it possible to use those touchscreens properly by adding
a "syna,no-pressure"+"syna,no-width" property for the DT and use
that to handle the data correctly.

Signed-off-by: Stephan Gerhold <stephan@gerhold.net>