arm64: dts: qcom: msm8916-ufi: make UDC dual mode
It's possible to use this device with a (non-standard) hub to get USB
working in host mode, but dr_mode="peripheral" prevents the UDC to
do so.
Remove dr_mode="peripheral"  and add usb-role-switch so that it defaults
to otg mode and can be switched to host mode in userspace.

Signed-off-by: Yang Xiwen <forbidden405@outlook.com>