power: supply: smb1360: Set hard-coded charging currents
The charging currents are not yet handled. Actually they should be set in
runtime based on connector type detection (like an extcon device but for
smb1360 a more complicated mechanism in USB PHY layer is needed, e.g. like
in downstream [1]). The upper limits for the charging currents should be read
from the devicetree.

As a temporary solution, set hard-coded charging currents in the
smb1360_hw_init() function.

[1] https://github.com/bq/aquaris-X5/blob/aquaris-X5_5.x/drivers/usb/phy/phy-msm-usb.c#L2959-L3112

Signed-off-by: Jakob Hauser <jahau@rocketmail.com>