arm64: dts: qcom: Increase smb1360 charging timeout
The currently implemented charging timeout of 192 minutes (approx. 3 hours)
turned out to be too short. When reaching the timeout, the device starts
to discharge even though the charging cable is still attached.

Therefore increase the charging timeout to 768 minutes (approx. 13 hours).
This is the default setting in the smb1360 chip and also the value of some of
the downstream devicetrees.

Disabling the charging timeout entirely was considered but discarded in
favor of safety. The timeout could avoid charging a faulty battery for too
long. On the other hand, the probability of reaching the timeout of 13 hours
by regular usage is considered very low – the timeout doesn't apply when
the battery is reaching full state and the charger switches to recharge cycle.

Signed-off-by: Jakob Hauser <jahau@rocketmail.com>